updating all requests to handle request counter functionality

+++ src/main/java/com/edafa/demo/controller/StudentController.java
@@ -5,11 +5,14 @@ import com.edafa.demo.entity.Student;
 import com.edafa.demo.service.CourseService;
 import com.edafa.demo.service.InstructorService;
 import com.edafa.demo.service.StudentService;
+import com.edafa.demo.user.UserRequestHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.web.bind.annotation.*;
 
 
+import java.lang.reflect.Array;
+import java.util.Arrays;
 import java.util.List;
 
 @RestController
@@ -21,27 +24,49 @@ public class StudentController {
 
     @GetMapping("/students")
     public List getAllStudents(){
-        return studentService.getAll();
+        if(UserRequestHandler.requestFlag)
+        {
+            return studentService.getAll();
+        }
+        return Arrays.asList("You exceeded your request limit");
     }
 
     @GetMapping("/student")
     public List getByNationality(@RequestParam String nat){
-        return studentService.getByNationality(nat);
+        if(UserRequestHandler.requestFlag)
+        {
+            return studentService.getByNationality(nat);
+        }
+        return Arrays.asList("You exceeded your request limit");
+
     }
 
     @PostMapping("/student")
     public String addNewStudent(@RequestBody Student st){
-        return studentService.registerNewStudent(st);
+        if(UserRequestHandler.requestFlag)
+        {
+            return studentService.registerNewStudent(st);
+        }
+        return ("You exceeded your request limit");
     }
 
     @PutMapping("/student")
     public String updateStudent(@RequestParam int id,@RequestBody Student st){
-        return studentService.updateStudent(id,st);
+        if(UserRequestHandler.requestFlag)
+        {
+            return studentService.updateStudent(id,st);
+        }
+        return ("You exceeded your request limit");
     }
 
     @DeleteMapping("/student")
     public String deleteStudent(@RequestParam int id){
-        return studentService.deleteStudent(id);
+        if(UserRequestHandler.requestFlag)
+        {
+            return studentService.deleteStudent(id);
+        }
+        return ("You exceeded your request limit");
+
     }
 
 }
